+++ src/components/CardList/CardList.jsx
@@ -1,7 +1,7 @@
 import * as S from "./CardList.style";
 import Card from "components/CardList/Card";
 import { Link } from "react-router-dom";
-import { Button } from "components/StyledButtons";
+import { Button } from "components/Button";
 
 function CardList({ items }) {
   const cards = items.recentMessages;
